+++ src/main/java/um/tesoreria/guarani/hexagonal/guarani/alumnos/alumno/infrastructure/web/controller/AlumnoGuaraniController.java
@@ -6,7 +6,6 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.server.ResponseStatusException;
@@ -58,6 +57,12 @@ public class AlumnoGuaraniController {
                 .toList());
     }
 
+    @GetMapping("/generate/personales/documento/{nroDocumento}")
+    public ResponseEntity<Boolean> createPersonalesByNroDocumento(@PathVariable String nroDocumento) {
+        log.debug("GET AlumnoGuaraniController.createPersonalesByNroDocumento for {}", nroDocumento);
+        return ResponseEntity.ok(service.createPersonalesByNroDocumento(nroDocumento));
+    }
+
     @GetMapping("/generate/preuniversitario/test")
     public ResponseEntity<Void> generatePreuniversitarioTest() {
         log.debug("\n\nProcessing AlumnoGuaraniController.generatePreuniversitarioTest\n\n");
@@ -65,11 +70,12 @@ public class AlumnoGuaraniController {
         return ResponseEntity.noContent().build();
     }
 
-    @PostMapping("/generate/preuniversitario/documento/{nroDocumento}")
+    @GetMapping("/generate/preuniversitario/documento/{nroDocumento}")
     public ResponseEntity<List<AlumnoGuaraniResponse>> createPreuniversitarioByNroDocumento(@PathVariable String nroDocumento) {
         log.debug("POST AlumnoGuaraniController.createPreuniversitarioByNroDocumento for {}", nroDocumento);
         return ResponseEntity.ok(service.createPreuniversitarioByNroDocumento(nroDocumento).stream()
                 .map(mapper::toResponse)
                 .toList());
     }
+
 }